requirements.txt
alembic==1.0.3
Babel==2.6.0
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dominate==2.3.4
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mtranslate==1.6
PyJWT==1.6.4
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
redis==3.0.1
requests==2.20.1
rq==0.13.0
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
